change the hypre solver.  From Weiqun:
The multi-D Rad2TShock tests have periodic boundaries in the directions
parallel to the shock surface.  The y and z velocities are supposed to be
zero.  With 103 or 1003, we do get perfect zero.  But with SMG and PFMG,
these velocity grow out of control.  I don't know whether that is a
problem of hypre's algorithm or our hydro or we did not set up periodic
boundary correctly in our radiation code interacting with hypre.  The
periodic case was never really used.